requirements.txt
asgiref==3.2.10
Django==3.0
pytz==2020.1
sqlparse==0.3.1
django-extensions==3.0.3
django-filter==2.3.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
importlib-metadata==1.7.0
jmespath==0.10.0
Markdown==3.2.2
Pillow==7.2.0
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.10
zipp==3.1.0
